backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
beanie>=1.26.0
pydantic[email]>=2.7.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
anthropic>=0.28.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1